RAM helper mappers table is now optional:
- The UNAPI specification now states that a RAM helper must provide a
  mappers table only if no mapper support routines are present in the system.

- The RAM helper installer v1.2 does exactly that: doesn't provide a mappers
  table if mapper support routines are present.

- The RAM and TST examples now rely on the RAM helper mappers table
  only if standard table is not present.

Introduced a new tool: MSR.COM, it installs DOS 2 compatible mapper support
routines in DOS 1, and installs the UNAPI RAM helper too. It will also
set the same slot for all pages of TPA if they aren't.
MSR.COM is generated from the same sources of the RAM helper
changing the value of a constant.

Changes in the UNAPI specification document:

- Providing a mappers table is now optional for RAM helpers when mapper
  support routines are present

- Included the mapper support routines API specification for reference